+++ HISTORY.md
@@ -14,6 +14,7 @@ Changes to JGraphT in each version:
     - Added `GraphSpecificsStrategy` and use the same edge set factory consistently (contributed by Dimitrios Michail)
     - Added user overview doc (contributed by John Sichi)
     - Clarified definition of `SimpleGraph` (contributed by Joris Kinable)
+    - Added O(m^1.5) algorithm for counting triangles in undirected graphs (contributed by Alexandru Văleanu)
 
 - **version 1.2.0** (16-May-2018):
     - Prepared release cycle 1.1.1: removed deprecated code, updated dependencies, etc (contributed by Joris Kinable)